Change CopyBufferToImage signature to match other methods
Take a pointer instead of a reference. Note: I agree that
it's better to take a reference to protect against nullptr, but
we should probably do that as a single mass refactor.

This also fixes a build problem in our internal build which
complains about {pCopyInfo} not actually constructing a vector
and ending up calling the enclosing function in an infinite loop.