Change CopyBufferToImage signature to match other methods (#38)
Take a pointer instead of a reference. Note: I agree that it's better to
take a reference to protect against `nullptr`, but we should probably do
that as a single mass refactor.

This also fixes a build problem in our internal build: `{pCopyInfo}` is
not actually constructing a vector (and the function is called in an
infinite loop).