grfx: fix issue with compressed textures
DXT* documentation states an image cannot have a size which
is not a multiple of the block-size.
However, it is possible to generate mipmap levels which size is less
then than minimum alignment, causing a validation error.

This commit tried to fix the loading by computing the correct stride,
and prevent the loading of a bogus texture.