fix: fix DX12 issue with compressed textures
For compressed textures like BC1, a block is 4x4 pixels.
Each block is 8 bytes, meaning a 4 pixel row is 2 bytes.
When computing the rowStride, I assumed we wanted the actual
byte stride:
- 2 bytes * width / 4 (as a block is 4 pixels large, and 2 bytes).

But trying that on DX12 fails as it expects a stride of 2 bytes * width.
My guess is it's not the actual file stride, but the file stride as if
each pixel was 2 byte width.
Not sure why, but it passes validation now, and runs fine on VK, DX12
and DX11.