Add a pack_glb utility to create glb files from GLTF (#47)
I have another script that uses this as a library as part of the process
of compressing all images and then packing them into a GLB, but this
works standalone as well. I split up the commits for PR readability and
because we haven't decided on exactly what extension and container
format to use for compressed images.