Rework Bitmap::Fill to avoid undefined behavior (#53)
Currently, `Bitmap::Fill` takes a float array and converts to integer
when needed. This is undefined behavior as the desired integer value may
not be representable in floating point.

This PR changes `Bitmap::Fill` to take a templated type instead. The
helper `CreateTexture1x1` is also changed to allow the caller to specify
the data type (which will be used to determine the type of the
underlying texture too).