benchmark: add ability to change shader on the fly
This is a first requirement for this benchmark: allow us to
have multiple shaders running and swap them at runtime, or not.

The PBR code is taken from the 28_gltf sample, but common parts are
moved to a shared file.
The idea is to share the same set of inputs across all shaders for
this benchmark, simplifying the pipeline layout handling.

Signed-off-by: Nathan Gauër <brioche@google.com>