Android support for BigWheels (#46)
This is a rough "as-little-change-as-possible" change to enable
BigWheels to run on Android. It has been tested with the gbuffer and
fishtornado samples successfully (Note: Fishtornado works on Quest2 but
fails on Pixel 6. We will have to re-visit fishtornado as a separate
task). Among some of the things changed to support Android:

- Disable imgui
- Disable cpu_features
- Abstract file loading so it goes through Android's asset manager (to
load assets from APK)
- Use memory-loading versions of 3rd party libraries (tinyobj, stbi,
etc) to avoid using freads
- Use Android surface and #ifdef out all the glfw code (including input
handling)
- Added all the necessary files to open and build BigWheels from
AndroidStudio via Gradle

Todo:
- Avoid populating root assets folder with build artifacts (i.e. SPV
files)
- Enable cpu_features
- Enable imgui for Android (not sure if it's practical to do this)
- Handle inputs
- Revisit landscape vs portrait orientation and render resolution
- Find/design some way to run different samples without editing
CMakelists
- Fix fishtornado so that it can run on a Pixel 6

---------

Co-authored-by: Shan-Min Chao <shanminchao@google.com>