Indirect lighting for PBR shader in basic_material
This change adds indirect lighting for the PBR shader in the
basic_material sample. For performance on mobile, the BRDF integration
is done using an approximation algorithm instead a LUT.

This change also contains irradiance and environment maps for the
indirect lighting. These maps were baked using a tool. The baking
process acounts for hemispherical sampling for the irradiance map and
roughness on the environment map. This can't be done at run time, it
would take forever to generate. The tool will be checked in in a
different PR.

- Added indirect lighting calculation to PBR shader.
- Changed Mipmap::LOadFile to load vertically tailed mip maps.
- Added method to creat texture from Mipmap.
- Added EnvDraw.hlsl for environment drawing with color correction and
  gamma adjustment.