android: switch back to NativeActivity.
NativeActivity is supported since a long time. GameActivity is the
new recommended method.
THe issue is GameActivity is not supported by Android, and being brand
new means lack of support in other domains.

Moving back to NativeActivity is a risk, if it suddenly gets deprecated.
But the cost of the switch being lower than the cost of fixing
dependencies to use GameActivity means it's an acceptable risk.

Signed-off-by: Nathan Gauër <brioche@google.com>