Switch mipmap generation to use box filter
This change switches the mipmap generation to use STBI's BOX filter
rather than the DEFAULT, which prefers quality. Since this application
does not measure visual fidelity, only performance, there's no reason to
be particularly picky when it comes to the quality of mipmaps generated
on the CPU at load time.