Enabled window resizing for Windows desktop (#144)
This change enables window resizing for Windows desktop applications for
D3D12 and Vulkan. Each API takes a different path for resizing due how
their swapchains are handled.

D3D12: `IDXGISwapChain::ResizeBuffers()` is used.
`grfx::Swapchain::Resize()` was added to handle window resize events.
There is a weird workaround for D3D12 to correctly invalidate the client
area on resizes.

Vulkan: The swapchain is destroyed and recreated.
`grfx::Swapchain::Resize()` returns an error on Vulkan.

Changes:
- Removed the surface caps caching for Vulkan. The caching was causing
the surface width/height to be locked into when the surface was created.
- Fixed AddRef/Releaase related bug in D3D12's image class that was
causing resource references not to be released when the resource was
external.
- Enabled window resizing for 01_triangle sample.
- Fixed some casting warnings

NOTE: Linux fixes will come in a later PR once we figure out the
situation with the ordering of resizing events relative to Acquire and
Present.