Minor adjustments to hysw's Android rotation
- Adds event handling for WINDOW_RESIZED on Android just to cover cases
  that may not be covered by CONTENT_RECT_CHANGED.
- Corrected behavior where in grfx::vk::Swapchain::PresentInternal that always
  assumed vkres != VK_SUCCESS was a failure. The behavior in
  grfx::vk::Swapchain::AcquireImageInternal is the desired behavior so
  present now just copies that.
- Added a comment regarding the situation with surface pretransform.
  Copied below for convenience.
- Suppressed swapchain warnings on Android for now.

Note about Vulkan surface pretransform:
Currently, we're going to assume IDENTITY for all platforms.
On Android we will incur the cost of the compositor doing the
rotation for us. We don't currently have the facilities in
place to inform the application of orientation changes
and supply it with the correct pretransform matrix.