Readjust positioning code to work properly
This change dives deeper into the positioning code and why it was
misbehaving. It turns out XR applications could adjust the position of
the ImGui UI by moving the quad onto which it was rendered. This isn't a
great approach in general as ImGui has its own concept of position
within the space its working, and moving the quad overly complicates
things.

With that working, XR apps now display diagnostics in the center of the
screen with automatic sizing, while desktop apps will spawn their window
initially at the top left of the screen, but obey the imgui.ini file
thereafter.