benchmark: add multiple sphere materials.
This commits adds the option to change the default model/scene
from the altimeter, to a sphere, with 6 LOD levels.
From 16k faces to 16.

To implement this, we could either swap meshes, or whole scenes.
The reason to swap scenes is to support complex scenes without adding
complexity.
Drawback is the skybox is not configurable at the scene-level, but can
be removed by adding a new scene like "sphere_0_no_skybox".

The materials are available in the GLTF files, but the current shaders
implementations don't switch to emissive/unlit the the skybox.

Signed-off-by: Nathan Gauër <brioche@google.com>