Fix debug output of fishtornado_xr (#162)
This change reenables the extended debug output of fishtornado_xr. To do
so, it tracks metrics for each view needed by the XR layer, and then
presents the sum (e.g. for GPU time, vertices, etc.). It also sets the
window position for the diagnostic output to anchor to windowWidth/4 and
windowHeight/4, with a pivot point at 0.5x0.5. This theoretically serves
to center the diagnostic output on XR as the imgui surface is the same
size as each eye, but spans across both. (This doesn't explain the
vertical difference well, though...)

---------

Co-authored-by: Lars Harrison <slumpwuffle@google.com>