utils: add ScopedTimer class.
A lot of timer patterns we have is start, do something, and stop
in the same function to measure some process.

The Scopedtimer wraps this logic in a class using RAII.
An alternative implementation could be to not print, but
use a float& which is set to the elapsed time in the destructor.
This would allow more control over the format, and print additional
info not available on timer initialization.
But for now, this is a simple pattern we can start with.

Signed-off-by: Nathan Gauër <brioche@google.com>