fix undefined behavior in Knobs
Knobs have this logic to raise an update flag on update.
When resetting the knob default value, the update flag is
only raised if the value changes.

The issue is during initialization: members have undefined values,
meaning for `if(currentValue == newValue)` or `if(updateNeeded)`,
results were dependent on uninitialized memory.

Signed-off-by: Nathan Gauër <brioche@google.com>