fs: tries to uniformize Android/Desktop FS ops. (#197)
Android has 2 ways to access assets:
 - the Assets NDK API
 - normal disk read. Desktop only has the latter.

The FS API built today was preventing Android to load assets from the
disk, and API diverged a bit (ex memory-mapped usage was Android only).

This commit tried to re-uniformize this. To limit the PR scope, the
memory mapping is not added for desktop, but is trivial to add now.

Adding this code would also unblock Android application to load assets
not from the APK folder, but from the disk. The issue I had was APK/disk
duality is only valid on desktop.
I solved this by assuming relative path on Android means in-APK, and
absolute means "I know what I'm doing, disk access".

This seems to work for most use-cases I found, but we might need to
revisit this in the future if need be.. Maybe we should stop try to
guess path between 3p and 1p assets, and maybe we could add some
`apk:/some/path` or `storage:/some/other/path` indicator? but that would
require apps to know in advance what we do with their assets..

---------

Signed-off-by: Nathan Gauër <brioche@google.com>