Move metrics report to be internal to manager
Due to the lifecycle of the data contained in a metrics report, it makes
the most sense to keep it associated with the manager (and consequently
its runs and metrics). It also makes sense to keep the lifetime of a
Report as short as can be, limiting its risk.

This change covers both of those, and also moves much of the logic for
report filename generation into the report itself. Finally, it
handles report output on Android correctly.