Add application-level metrics management wrappers.
This change introduces application-level metrics management functions,
which avoids the use-after-free pure-pointer approach to metric
exposure. Instead, this applies knowledge of the application's use of
metrics to create a lookup table of metric IDs to pointers as necessary,
and introduces a novel functional interface to record metrics.

Basically, this wraps the metrics functionality at the application
level.