samples: fix asset paths loading for some samples
LoadShader take a subdir, and a shader name. Those 2 are then
combined, and GetAssetPath is called.
2 Samples however called GetAssetPath on the subdir before.
This was fine as long as the path returned was absolute because
std::filesystem::path ignores the first part when
concatenating to an absolute path:

```
"/tmp/titi/" / "/tmp/toto" == "/tmp/toto"
"/tmp/titi/" / "tmp/toto" == "/tmp/titi/tmp/toto"
```

Now what we use "--assets-path", the asset dir could be relative, and
this bugs becomes visible.

Signed-off-by: Nathan Gauër <brioche@google.com>