Fixes view offset sizes in D3D12 (#231)
D3D12 views for buffer objects requires a size to be specified. This is
analogous to Vulkan's range for storage and uniform buffers. Previous to
this commit, the D3D12 views always assumed the size of the buffer for
index and buffer views. Constant buffer view offset and size were
already handled. correctly. This change corrects handling of both
offsets and sizes for index and vertex buffer views.

- Added size member variable to IndexBufferView and VertexBufferView.
- Fixed missing offset in index and vertex buffer views for D3D12.
- Fixed missing size in index and vertex buffer view for D3D12.
- Removed unused code.