Overhaul metrics
This change massively updates the way that metrics work from exposing
pointers externally to using an ID. This forces metric updates to go
through the manager, but alleviates concerns about pointer ownership and
expiration.

As part of this, the metrics themselves have been generalized using the
'Metric' abstract class, which reduces code duplication. To accomplish
this, some of the existing functionality was eliminated, but this did
not appear to be used anywhere but in tests.

The tests, likewise, have been updated to instead prefer the eventual
output (the Report) as validation, rather than any internal knowledge.