Update CLI argument naming and help spacing
This change alters the spacing for --help output to be 20 characters for
descriptions. This allows a good amount more text per line, but means
longer arguments now have line separation between the argument and the
description. Fortunately, this has the effect of looking pretty
homogeneous with any length of argument, and allows --xr-ui-res to
return to --xr-ui-resolution.