Scene Graph + Renderer Part 1 (#304)
***NOTE: ERROR_SCENE_* enums aren't final yet**

This change adds an initial implmentation of a scene graph. Each object
has documentation and usage guidelines. GLTF loader and examples will
follow in next PR.

Primary TL;DR;s:
- Implementation uses std::shared_ptr for shared object referencing -
for better or for worse.
- Some scene graph objects (images, samplers, textures, materials,
meshes, and nodes) can be used in standalone mode and doesn't have to be
apart of a scene. This is useful for when someone wants to use GLTFs as
an asset library instead of a scene.