geometry: change resize policy to be double (#312)
The initial resize policy for the buffer was "minimalist". Only resized
to match the added element.

Changed this policy to be "double the size". Every time we reach the
size limit, the underlying container size is doubled.

This reduces the benchmark loading time from 30s to 15s in debug. (367M
resize calls to 725).

---------

Signed-off-by: Nathan Gauër <brioche@google.com>