A few changes to XR handling
Swapchain:
- Add Wait() method to swapchain, it is a no-op for non-XR. For XR it
calls xrWaitSwapchainImage. Since the command buffer can be recorded
between xrAcquireSwapchainImage and xrWaitSwapchainImage, putting
xrWaitSwapchainImage in AcquireNextImage is unnecessarily restrictive.
- Add xrReleaseSwapchainImage to Present(). It is weird not to have a
function for this and to do this in the main application loop, since the
application already calls Acquire.

Remove calling Render() twice for XR and the viewIndex logic. It is very
much application/project specific logic of how it decides to render two
views, and can be easily handled in the application level. It also
becomes more restrictive where application decides to render differently
(for example, using a single command buffer for UI and both views, or
with multiview