Not for submissiong, some thought experiments
- There is a bit of a weird relationship between settings and default values
for things that repeat: resolution and XR UI resolution. If the defaults
for those are provided then the values in the settings will be overwritten.
Might not be a problem, but might not be immediately obvious.

- There is an ordering dependencies between some functions in the beginning
of Run(). They seem to be documented, but comments don't seem to be the
best way for this, as some comments already might cause confusion. This
explores the option of providing defaults in the InitKnobs instead of as
part of the settings.

- Removing the storage for the flags in command line parser, this seems to be
only needed for `--help`, so it shouldn't be necessary to store that.