Fix invalid timestamp computation for single-pass rendering (#375)
The time computation for multi-pass rendering was incorrect. For both
`Timestamp Queries` to the GPU, the PIPELINE_STAGE_TOP_OF_PIPE_BIT stage
was erroneously used. To correct this, the first timestamp should use
the PIPELINE_STAGE_TOP_BIT stage to record the timestamp as soon as all
previous commands are processed by the command processor. The second
timestamp should employ the PIPELINE_STAGE_BOTTOM_BIT stage to record
the timestamp as soon as all previous commands retire.