Flatten ComputeShader hierarchy
This removes the previous hierarchy of compute shaders by commoning out
the dispatch interface.  The basic observation here is that all the
compute shaders differ only in their input signatures.

This makes the signature part of shader construction.  When the shader
is created, a vector of integers specifies what binding slots to use for
the input textures and the order in which they need to be presented.

Additionally, all the scalar data needed during invocation is passed
explicitly via the ScalarInput structure (designated initializers would
be great here, btw).

Other changes:

- Update marble drop frequency and splat frequency defaults to be less
  visually noisy.

- Remove the unnecessary color grid shader (artifact from the original
  JavaScript implementation).

- Fix #230. The original code was not setting `texelSize` appropriately..