Re-factor compute and rendering support (#396)
This simplifies the logic handling scheduling and dispatch of compute
shaders and textures.

Main changes:

- Class Texture renamed to SimulationGrid for clarity. The supporting
textures are created from Bitmaps. Each grid has its own descriptor set
and vertex buffer.

- Simplified class hierarchy for shaders. This removes the
GraphicsShader class and moves the rendering code to
SimulationGrid::Draw().

- Removed the concept of dispatch records. Instead, each compute shader
has a Dispatch() method that builds up (or re-uses) the descriptor set
with all the arguments and queues into the frame's command buffer. This
fixes #26

- Removed structs `ComputeResources` and `GraphicsResources`. Moved
members closer to where they're needed inside `FluidSimulationApp` or
`SimulationGrid`.

- Renamed `FluidSimulation` to `FluidSimulationApp` and made it inherit
from `ppx::Application`. Moved all the application code to `sim.cpp`.
This removes the need to have a simulation object.

- Compute dispatches are all grouped together before the render pass.

- Data required to be used in each dispatch operation is grouped inside
the PerFrame structure. This is implemented as a pool that contains
descriptors and uniform buffers to use for compute dispatch operations.
On every frame, every entry in the pool is marked available. When
scheduling a new compute operation, if there are not enough entries in
the pool, a new one is added (in PerFrame::GetDispatchData).

- Removed unused code to generate and display a checkerboard grid
(remnant from original port).

A subsequent cleanup will simplify the compute shader hierarchy. It's
now largely unnecessary, but this change was becoming too large already.