Wait for render complete before image acquisition
Recent validation layers flag that imageAcquisitionSemaphore may still
be signalled during AcquireNextImage. The only logic between the bottom
of Render for the previous frame and the top of Render for the next
frame is the Application frame rate pacing; there is no explicit
synchronization. As such, applications with no frame pacing or those
that cannot hit the target FPS may not have completely finished
rendering the previous frame when it requests to render the next frame.

The fix is to wait for renderCompleteFence before AcquireNextImage so
that we guarantee that imageAcquisitionSemaphore is unsignalled.