Add UINT8 index buffer support to Geometry (#515)
Add support for creating buffers that the GPU can accept if
IndexTypeUint8Supported(). The client must decide which index type to
use depending on the supported features of the API and GPU; it is
possible to create a Uint8 buffer which i.e. is not accepted by DX12.

This is required for #453