[glTF] Load scenes with UINT8 indices (#482)
Load scenes that contain 8-bit index buffers. How the index buffer ends
up being packed depends on the capabilities of the device:

- On Vulkan with `VK_KHR_index_type_uint8` extension, the UINT8 indices
can be packed into a `INDEX_TYPE_UINT8` buffer. This is the optimal case
as it doesn't require additional memory.
- Otherwise, the UINT8 indices need to be repacked into UINT16 indices.
This comes at the cost of a larger GPU buffer.

This allows us to load more Khronos
[glTF-Sample-Assets](https://github.com/KhronosGroup/glTF-Sample-Assets)
like
[TextureCoordinateTest](https://github.com/KhronosGroup/glTF-Sample-Assets/tree/main/Models/TextureCoordinateTest).


![image](https://github.com/user-attachments/assets/70985607-ed30-4354-9991-436187db4337)

Fixes #453