Upgrade Linux job runner to Ubuntu 24.04 (#556)
Ubuntu 20.04 has been removed as a runner, causing checks to fail on
PRs.

Ubuntu 24.04 was chosen in the hopes that we won't have to think about
it for a few more years than if we were to choose Ubuntu 22.04.

This required additional changes:

- Linux build requires `convert` so make sure imagemagick is installed.
- clang-format was pinned to `clang-format-12` but this package is not
in 24.04. The oldest version that 24.04 offers is `clang-format-14`. I'm
worried about pushing this too high, having it change how it wants code
formatted, and creating churn for PR authors. Though maybe I'm
overthinking this.

Fixes #555