Reapply "Safely reuse Present wait semaphore" (#571) (#578)
This reverts commit 7add2d9e40f31377afcebc3380c0b2be7659e1f3.

PR #565 introduced a stack-use-after-scope bug. This was reverted by
#571.

While the compiler normally prevents taking the address of an rvalue,
SemaphorePtr (ObjPtr<Semaphore>) overloads unary `operator&`; it appears
that this overload is not scrutinized well by the compiler since it did
not include a ref-qualifier. A ref-qualifier was added in #579

Additionally, to address this issue, Semaphore* was used instead of
SemaphorePtr. See #572 for more info.

Finally, all projects have been manually updated to use
`GetPresentationReadySemaphore()`. This was a good move anyway since
some samples were incorrectly fixed the first time (like fishtornado).

Tested by:
- running internal presubmits
- building with SDK 1.4.321 and running fishtornado Vulkan on Linux,
fishtornado Vulkan and DirectX on Windows